refactor: search input generic component
