+++ mobile-app/app/components/HeaderSearchInput.tsx
@@ -1,0 +1,51 @@
+import { View } from '@components'
+import { tailwind } from '@tailwind'
+import { translate } from '@translations'
+import { Platform, TouchableOpacity } from 'react-native'
+import { useSafeAreaInsets } from 'react-native-safe-area-context'
+import { SearchInput } from './SearchInput'
+import { ThemedIcon, ThemedView } from './themed'
+import React from 'react'
+
+interface HeaderSearchInputProps {
+  searchString: string
+  onClearInput: () => void
+  onChangeInput: (text: string) => void
+  onCancelPress: () => void
+}
+
+export function HeaderSearchInput (props: HeaderSearchInputProps): JSX.Element {
+  const safeAreaInsets = useSafeAreaInsets()
+  return (
+    <ThemedView
+      light={tailwind('bg-white border-gray-200')}
+      dark={tailwind('bg-gray-800 border-gray-700')}
+      style={[
+        tailwind('flex flex-row items-center pb-2 px-4'),
+        {
+          paddingTop: Platform.OS === 'android' ? safeAreaInsets.top + 8 : safeAreaInsets.top - 4
+        }
+      ]}
+    >
+      <SearchInput
+        value={props.searchString}
+        placeholder={translate('screens/ChooseLoanTokenScreen', 'Search for loans')}
+        autoFocus
+        showClearButton={props.searchString !== ''}
+        onClearInput={props.onClearInput}
+        onChangeText={props.onChangeInput}
+      />
+      <View style={tailwind('flex justify-center ml-2')}>
+        <TouchableOpacity onPress={props.onCancelPress}>
+          <ThemedIcon
+            light={tailwind('text-primary-500')}
+            dark={tailwind('text-darkprimary-500')}
+            iconType='MaterialCommunityIcons'
+            name='close'
+            size={24}
+          />
+        </TouchableOpacity>
+      </View>
+    </ThemedView>
+  )
+}
