+++ mobile-app/app/screens/AppNavigator/screens/Loans/screens/ChooseLoanTokenScreen.tsx
@@ -1,19 +1,17 @@
-import { ThemedIcon, ThemedView } from '@components/themed'
+import { ThemedView } from '@components/themed'
 import React, { useCallback, useEffect, useLayoutEffect, useState } from 'react'
 import { LoanCards } from '../components/LoanCards'
 import { StackScreenProps } from '@react-navigation/stack'
 import { LoanParamList } from '../LoansNavigator'
 import { debounce } from 'lodash'
-import { Platform, TouchableOpacity, View } from 'react-native'
 import { tailwind } from '@tailwind'
-import { translate } from '@translations'
-import { useSafeAreaInsets } from 'react-native-safe-area-context'
-import { SearchInput } from '@components/SearchInput'
 import { useDispatch, useSelector } from 'react-redux'
 import { RootState } from '@store'
 import { LoanToken } from '@defichain/whale-api-client/dist/api/loan'
 import { useWhaleApiClient } from '@shared-contexts/WhaleContext'
 import { fetchLoanTokens } from '@store/loans'
+import { HeaderSearchIcon } from '@components/HeaderSearchIcon'
+import { HeaderSearchInput } from '@components/HeaderSearchInput'
 
 type Props = StackScreenProps<LoanParamList, 'ChooseLoanTokenScreen'>
 
@@ -77,66 +75,6 @@ export function ChooseLoanTokenScreen ({ navigation, route }: Props): JSX.Elemen
         loans={filteredLoans}
         vaultId={vaultId}
       />
-    </ThemedView>
-  )
-}
-
-function HeaderSearchIcon (props: {onPress: () => void}): JSX.Element {
-  return (
-    <TouchableOpacity
-      onPress={props.onPress}
-      style={tailwind('pr-4')}
-    >
-      <ThemedIcon
-        light={tailwind('text-primary-500')}
-        dark={tailwind('text-darkprimary-500')}
-        iconType='MaterialIcons'
-        name='search'
-        size={24}
-      />
-    </TouchableOpacity>
-  )
-}
-
-interface HeaderSearchInputProps {
-  searchString: string
-  onClearInput: () => void
-  onChangeInput: (text: string) => void
-  onCancelPress: () => void
-}
-
-function HeaderSearchInput (props: HeaderSearchInputProps): JSX.Element {
-  const safeAreaInsets = useSafeAreaInsets()
-  return (
-    <ThemedView
-      light={tailwind('bg-white border-gray-200')}
-      dark={tailwind('bg-gray-800 border-gray-700')}
-      style={[
-        tailwind('flex flex-row items-center pb-2 px-4'),
-        {
-          paddingTop: Platform.OS === 'android' ? safeAreaInsets.top + 8 : safeAreaInsets.top - 4
-        }
-      ]}
-    >
-      <SearchInput
-        value={props.searchString}
-        placeholder={translate('screens/ChooseLoanTokenScreen', 'Search for loans')}
-        autoFocus
-        showClearButton={props.searchString !== ''}
-        onClearInput={props.onClearInput}
-        onChangeText={props.onChangeInput}
-      />
-      <View style={tailwind('flex justify-center ml-2')}>
-        <TouchableOpacity onPress={props.onCancelPress}>
-          <ThemedIcon
-            light={tailwind('text-primary-500')}
-            dark={tailwind('text-darkprimary-500')}
-            iconType='MaterialCommunityIcons'
-            name='close'
-            size={24}
-          />
-        </TouchableOpacity>
-      </View>
     </ThemedView>
   )
 }
